页面路由修改和部分代码优化

+++ API/maku-boot-travel/src/main/java/net/maku/travel/controller/TPlanTripMainController.java
@@ -2,19 +2,21 @@ package net.maku.travel.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
+import net.maku.framework.common.constant.Constant;
 import net.maku.framework.common.utils.PageResult;
 import net.maku.framework.common.utils.Result;
+import net.maku.framework.security.user.SecurityUser;
+import net.maku.framework.security.user.UserDetail;
 import net.maku.travel.convert.TPlanTripMainConvert;
 import net.maku.travel.entity.TPlanTripMainEntity;
-import net.maku.travel.service.TPlanTripMainService;
 import net.maku.travel.query.TPlanTripMainQuery;
+import net.maku.travel.service.TPlanTripMainService;
 import net.maku.travel.vo.TPlanTripMainVO;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-
-import jakarta.validation.Valid;
 
 import java.util.List;
 
@@ -35,8 +37,18 @@ public class TPlanTripMainController {
     @Operation(summary = "分页")
     @PreAuthorize("hasAuthority('travel:planMain:page')")
     public Result<PageResult<TPlanTripMainVO>> page(@ParameterObject @Valid TPlanTripMainQuery query) {
+        UserDetail user = SecurityUser.getUser();
+        if (!user.getSuperAdmin().equals(Constant.SUPER_ADMIN)) {
+            query.setCreator(user.getId().intValue());
+        }
         PageResult<TPlanTripMainVO> page = tPlanTripMainService.page(query);
+        return Result.ok(page);
+    }
 
+    @GetMapping("query")
+    @Operation(summary = "查询列表对外提供API")
+    public Result<PageResult<TPlanTripMainVO>> query(@ParameterObject @Valid TPlanTripMainQuery query) {
+        PageResult<TPlanTripMainVO> page = tPlanTripMainService.page(query);
         return Result.ok(page);
     }
 
@@ -54,6 +66,10 @@ public class TPlanTripMainController {
     @PreAuthorize("hasAuthority('travel:planMain:save')")
     public Result<String> save(@RequestBody TPlanTripMainVO vo) {
         vo.setState(0);
+        UserDetail user = SecurityUser.getUser();
+        if (!user.getSuperAdmin().equals(Constant.SUPER_ADMIN)) {
+            vo.setCreator(user.getId().intValue());
+        }
         tPlanTripMainService.save(vo);
 
         return Result.ok();
@@ -63,6 +79,10 @@ public class TPlanTripMainController {
     @Operation(summary = "修改")
     @PreAuthorize("hasAuthority('travel:planMain:update')")
     public Result<String> update(@RequestBody @Valid TPlanTripMainVO vo) {
+        UserDetail user = SecurityUser.getUser();
+        if (!user.getSuperAdmin().equals(Constant.SUPER_ADMIN)) {
+            vo.setCreator(user.getId().intValue());
+        }
         tPlanTripMainService.update(vo);
 
         return Result.ok();
@@ -72,7 +92,10 @@ public class TPlanTripMainController {
     @Operation(summary = "删除")
     @PreAuthorize("hasAuthority('travel:planMain:delete')")
     public Result<String> delete(@RequestBody List<Long> idList) {
-        tPlanTripMainService.delete(idList);
+        UserDetail user = SecurityUser.getUser();
+        if (user.getSuperAdmin().equals(Constant.SUPER_ADMIN)) {
+            tPlanTripMainService.delete(idList);
+        }
 
         return Result.ok();
     }
